fix: use multiple ghost ingredient handlers instead
